+++ app/events/karma.py
@@ -1,20 +1,17 @@
-from discord.ext import commands, tasks
-import discord
-from sqlalchemy import select
-
-from app.db.karma_db import Database, RewardsTable
 import logging
 
-class Karma(commands.Cog):
-    logging.basicConfig(level=logging.ERROR,
-                    format='%(asctime)s %(message)s',
-                    handlers=[logging.StreamHandler()])
+import discord
+from discord.ext import commands, tasks
+from sqlalchemy import select
 
+from db.tables import RewardsTable
+from db.user_karma import UserKarma
+
+class Karma(commands.Cog):
+    logging.basicConfig(level=logging.ERROR, format='%(asctime)s %(message)s', handlers=[logging.StreamHandler()])
 
     def __init__(self, bot):
         self.bot = bot
-        self.db = Database()
-        self.bot.loop.create_task(self.db.init_db())
         self.give_voice_karma.start()
 
     def cog_unload(self):
@@ -22,48 +19,39 @@ class Karma(commands.Cog):
 
     @discord.Cog.listener()
     async def on_guild_join(self, guild):
-            for member in guild.members:
-                if not member.bot:
-                    await self.db.create_user_record_in_karma(member.id, guild.id)
+        for member in guild.members:
+            if not member.bot:
+                await UserKarma().create_user_record_in_karma(member.id, guild.id)
 
     @discord.Cog.listener()
     async def on_message(self, message: discord.Message):
-        if (
-            message.author.bot or 
-            message.channel.id == 1229062537954332782 or # commands channel
-            message.channel.id == 1337733289695514725 or # counting channel
-            message.channel.id == 1339010562964586647 or # cult leader channel
-            message.channel.id == 1283842433284837396 # burgeramt channel
-        ): 
+        if (message.author.bot or message.channel.id == 1229062537954332782 or  # commands channel
+                message.channel.id == 1337733289695514725 or  # counting channel
+                message.channel.id == 1339010562964586647 or  # cult leader channel
+                message.channel.id == 1283842433284837396  # burgeramt channel
+        ):
             return
 
-        await self.db.handle_message_karma(
-            user_id=message.author.id,
-            guild_id=message.guild.id,
-            timestamp=message.created_at.timestamp(),
-        )
+        await UserKarma().handle_message_karma(user_id=message.author.id, guild_id=message.guild.id,
+            timestamp=message.created_at.timestamp(), )
 
     @tasks.loop(minutes=1)
     async def give_voice_karma(self):
         try:
-            async with self.db.get_session() as session:
-                for guild in self.bot.guilds:
-                    for channel in guild.voice_channels:
-                        active_users = [
-                            member for member in channel.members
-                            if not member.bot and not member.voice.self_mute and not member.voice.self_deaf
-                        ]
-                        if len(active_users) >= 2:
-                            for user in active_users:
-                                await self.db.adjust_karma_for_user(user.id, guild.id, amount=1)
-                await session.commit()
+            for guild in self.bot.guilds:
+                for channel in guild.voice_channels:
+                    active_users = [member for member in channel.members if
+                        not member.bot and not member.voice.self_mute and not member.voice.self_deaf]
+                    if len(active_users) >= 2:
+                        for user in active_users:
+                            await UserKarma().adjust_karma_for_user(user.id, guild.id, amount=1)
         except Exception as e:
             logging.error(f"Error during voice karma loop: {e}")
 
     async def manage_karma_rewards(self, guild_id, user_id):
         try:
-            rewards = await self.db.list_rewards(guild_id)
-            user_karma = await self.db.get_user_karma(user_id, guild_id)
+            rewards = await UserKarma().list_rewards(guild_id)
+            user_karma = await UserKarma().get_user_karma(user_id, guild_id)
 
             if user_karma is None:
                 logging.warning(f"No karma found for user_id={user_id} in guild_id={guild_id}")
@@ -99,12 +87,8 @@ class Karma(commands.Cog):
         if message.author.bot:
             return
 
-        await self.db.handle_reaction_change(
-            message_author=message.author,
-            guild_id=payload.guild_id,
-            emoji_id=payload.emoji.id,
-            is_addition=True
-        )
+        await self.db.handle_reaction_change(message_author=message.author, guild_id=payload.guild_id,
+            emoji_id=payload.emoji.id, is_addition=True)
 
     @discord.Cog.listener()
     async def on_raw_reaction_remove(self, payload):
@@ -115,12 +99,8 @@ class Karma(commands.Cog):
         if message.author.bot:
             return
 
-        await self.db.handle_reaction_change(
-            message_author=message.author,
-            guild_id=payload.guild_id,
-            emoji_id=payload.emoji.id,
-            is_addition=False
-        )
+        await self.db.handle_reaction_change(message_author=message.author, guild_id=payload.guild_id,
+            emoji_id=payload.emoji.id, is_addition=False)
 
     @discord.slash_command(name="adjustkarma", contexts={discord.InteractionContextType.guild})
     @discord.ext.commands.has_guild_permissions(administrator=True)
@@ -131,7 +111,7 @@ class Karma(commands.Cog):
         if member.bot:
             return await ctx.respond("Bots cannot receive karma!")
 
-        await self.db.adjust_karma_for_user(member.id, ctx.guild.id, amount)
+        await UserKarma().adjust_karma_for_user(member.id, ctx.guild.id, amount)
         await ctx.respond(f"Gave {amount} karma to {member.mention}!")
 
     @discord.slash_command(name="leaderboard", contexts={discord.InteractionContextType.guild})
@@ -139,7 +119,7 @@ class Karma(commands.Cog):
         """Displays the leaderboard for the server."""
         await ctx.defer()
         embed = discord.Embed(title="Karma Leaderboard", color=discord.Color.blurple())
-        top_users = await self.db.get_karma_leaderboard(ctx.guild.id, 10)
+        top_users = await UserKarma().get_karma_leaderboard(ctx.guild.id, 10)
 
         if not top_users:
             await ctx.respond("No leaderboard data available.")
@@ -155,14 +135,14 @@ class Karma(commands.Cog):
     @discord.ext.commands.has_guild_permissions(administrator=True)
     async def clear_leaderboard(self, ctx):
         """Clears the leaderboard for the server."""
-        await self.db.clear_karma_leaderboard(ctx.guild.id)
+        await UserKarma().clear_karma_leaderboard(ctx.guild.id)
         await ctx.respond("Leaderboard has been cleared!")
 
     @discord.slash_command(name="karma", contexts={discord.InteractionContextType.guild})
     async def check_karma(self, ctx, member: discord.Member = None):
         """Check karma for a user."""
         member = member or ctx.author
-        karma = await self.db.get_user_karma(member.id, ctx.guild.id)
+        karma = await UserKarma().get_user_karma(member.id, ctx.guild.id)
         await ctx.respond(f"{member.display_name} has {karma or 0} karma.")
 
     @discord.slash_command(name="add_reward", contexts={discord.InteractionContextType.guild})
@@ -178,25 +158,21 @@ class Karma(commands.Cog):
     @discord.ext.commands.has_guild_permissions(manage_roles=True)
     async def remove_reward(self, ctx, role: discord.Role):
         """Remove a reward role for karma."""
-        await self.db.remove_reward(role.id, ctx.guild.id)
+        await UserKarma().remove_reward(role.id, ctx.guild.id)
         await ctx.respond(f"Removed {role.name} from the reward roles.")
 
     @discord.slash_command(name="rewards", contexts={discord.InteractionContextType.guild})
     async def list_rewards(self, ctx):
         """List all reward roles for karma."""
         async with self.db.get_session() as session:
-            results = await session.execute(
-                select(RewardsTable).filter_by(guild_id=ctx.guild.id)
-            )
+            results = await session.execute(select(RewardsTable).filter_by(guild_id=ctx.guild.id))
             rewards = results.scalars().all()
         if not rewards:
             await ctx.respond("No rewards have been set.")
         else:
             rewards_list = "\n".join(
-                [f"{ctx.guild.get_role(reward.role_id).name}: {reward.karma_needed} karma" for reward in rewards]
-            )
+                [f"{ctx.guild.get_role(reward.role_id).name}: {reward.karma_needed} karma" for reward in rewards])
             await ctx.respond(f"Reward roles:\n{rewards_list}")
-
 
     @remove_reward.error
     @add_reward.error
@@ -204,36 +180,19 @@ class Karma(commands.Cog):
     @give_karma.error
     async def on_command_error(self, ctx, error):
         """Handles errors """
-        embed = discord.Embed(
-            title="Error",
-            description="Something went wrong while executing the command.",
-            color=discord.Color.red()
-        )
+        embed = discord.Embed(title="Error", description="Something went wrong while executing the command.",
+            color=discord.Color.red())
 
         if isinstance(error, commands.MissingPermissions):
-            embed.add_field(
-                name="Permission Denied",
-                value="You lack the required permissions to run this command.",
-                inline=False
-            )
+            embed.add_field(name="Permission Denied", value="You lack the required permissions to run this command.",
+                inline=False)
         elif isinstance(error, commands.MemberNotFound):
-            embed.add_field(
-                name="Member Not Found",
-                value="The specified member could not be found.",
-                inline=False
-            )
+            embed.add_field(name="Member Not Found", value="The specified member could not be found.", inline=False)
         elif isinstance(error, commands.CommandInvokeError):
-            embed.add_field(
-                name="Command Error",
-                value=str(error.original),
-                inline=False
-            )
+            embed.add_field(name="Command Error", value=str(error.original), inline=False)
         else:
-            embed.add_field(
-                name="Unknown Error",
-                value="An unexpected error occurred. Please contact an admin.",
-                inline=False
-            )
+            embed.add_field(name="Unknown Error", value="An unexpected error occurred. Please contact an admin.",
+                inline=False)
 
         try:
             await ctx.respond(embed=embed, ephemeral=True)
