+++ db/user_karma.py
@@ -1,76 +1,13 @@
 import logging
-import os
-from contextlib import asynccontextmanager
-from time import time
 
-from dotenv import load_dotenv
-from sqlalchemy import Column, Integer, BigInteger, insert, select, delete
-from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker, scoped_session
+from sqlalchemy import insert, select, delete
 
-load_dotenv()
-Base = declarative_base()
-logging.basicConfig(level=logging.ERROR, format='%(asctime)s %(message)s', handlers=[logging.StreamHandler()])
+from db import Database
+from db.tables import KarmaTable, RewardsTable
 
 
-class KarmaTable(Base):
-    __tablename__ = 'karma'
-    user_id = Column(BigInteger, primary_key=True)
-    guild_id = Column(BigInteger, primary_key=True)
-    karma = Column(Integer, default=0)
-    timestamp_last_message = Column(Integer, default=0, )
+class UserKarma(Database):
 
-    def __init__(self, user_id, guild_id, karma, timestamp_last_message=time()):
-        super().__init__()
-        self.user_id = user_id
-        self.guild_id = guild_id
-        self.karma = karma
-        self.timestamp_last_message = timestamp_last_message
-
-    def dump(self):
-        return dict([(k, v) for k, v in vars(self).items() if not k.startswith('_')])
-
-
-class RewardsTable(Base):
-    __tablename__ = "rewards"
-    role_id = Column(BigInteger, primary_key=True)
-    guild_id = Column(BigInteger, primary_key=True)
-    karma_needed = Column(Integer, nullable=False)
-
-    def __init__(self, role_id, guild_id, karma_needed):
-        super().__init__()
-        self.role_id = role_id
-        self.guild_id = guild_id
-        self.karma_needed = karma_needed
-
-    def dump(self):
-        return dict([(k, v) for k, v in vars(self).items() if not k.startswith('_')])
-
-
-class Database:
-
-    def __init__(self):
-        db_url = f"mysql+pymysql://{os.getenv('DB_USER')}:{os.getenv('DB_PW')}@{os.getenv('DB_HOST')}:{os.getenv('DB_PORT')}/{os.getenv('DB_NAME')}"
-        self.engine = create_async_engine(db_url, pool_recycle=3600, pool_pre_ping=True, pool_use_lifo=True)
-        self.SessionLocal = scoped_session(
-            sessionmaker(engine=self.engine, class_=AsyncSession, expire_on_commit=False))
-
-    async def init_db(self):
-        """Create all tables in the db."""
-        async with self.engine.begin() as conn:
-            await conn.run_sync(Base.metadata.create_all)
-
-    @asynccontextmanager
-    async def get_session(self):
-        """Provide an async session."""
-        async with self.SessionLocal() as session:
-            try:
-                yield session
-            finally:
-                await session.close()
-
-    #############################################################################################################
     async def create_user_record_in_karma(self, user_id, guild_id):
         async with self.get_session() as session:
             existing_record = await session.execute(
